# Add search test to start profiling and investigate how to optimize search endpoint