--- a/src/flick/asset/serializers.py
+++ b/src/flick/asset/serializers.py
@@ -1,45 +1,21 @@
-from django.contrib.auth.models import User
-
-from rest_framework.serializers import CurrentUserDefault, ModelSerializer, PrimaryKeyRelatedField
+from rest_framework.serializers import CurrentUserDefault
+from rest_framework.serializers import ModelSerializer
 
 from .models import AssetBundle
 
 
 class AssetBundleSerializer(ModelSerializer):
-
-    # owner = PrimaryKeyRelatedField(read_only=True, default=CurrentUserDefault())
-
     class Meta:
         model = AssetBundle
-        fields = (
-            "id",
-            "salt",
-            "kind",
-            "base_url",
-            # 'owner',
-            "asset_urls",
-            "created_at",
-        )
-        read_only_fields = ("id",)
+        fields = ("id", "salt", "kind", "base_url", "asset_urls", "created_at")
+        read_only_fields = fields
 
     def save(self):
         self.user = CurrentUserDefault()
 
 
 class AssetBundleDetailSerializer(ModelSerializer):
-
-    # owner = UserSerializer(many=False, read_only=True)
-
     class Meta:
         model = AssetBundle
-        fields = (
-            "id",
-            "salt",
-            "kind",
-            "base_url",
-            # 'owner',
-            "asset_urls",
-            "created_at",
-            "updated_at",
-        )
-        read_only_fields = ("id",)
+        fields = ("id", "salt", "kind", "base_url", "asset_urls", "created_at", "updated_at")
+        read_only_fields = fields