--- a/src/flick/discover/views.py
+++ b/src/flick/discover/views.py
@@ -1,7 +1,7 @@
 from api.utils import success_response
 from django.core.cache import caches
 from rest_framework.views import APIView
-from show.api_utils import API
+from show.show_api_utils import ShowAPI
 from tag.models import Tag
 from tag.serializers import TagSerializer
 
@@ -15,7 +15,7 @@ class DiscoverShow(APIView):
     def get_top_shows_by_type(self, show_type):
         top_shows = local_cache.get(("top", show_type))
         if not top_shows:
-            top_shows = API.get_top_show_info(show_type)
+            top_shows = ShowAPI.get_top_show_info(show_type)
             local_cache.set(("top", show_type), top_shows)
         self.shows.extend(top_shows)
 
@@ -31,8 +31,8 @@ class DiscoverShow(APIView):
         top_shows = local_cache.get(("top", show_type, tag_id))
         if not top_shows:
             tag_data = TagSerializer(Tag.objects.get(id=tag_id)).data
-            ext_api_tag = tag_data.get("ext_api_id")
-            top_shows = API.get_top_show_info_by_genre(show_type, ext_api_tag)
+            ext_api_genre_id = tag_data.get("ext_api_genre_id")
+            top_shows = ShowAPI.get_top_show_info_by_genre(show_type, ext_api_genre_id)
             local_cache.set(("top", show_type, tag_id), top_shows)
 
         self.shows.extend(top_shows)
@@ -63,6 +63,6 @@ class DiscoverShow(APIView):
             self.get_top_shows(is_movie, is_tv, is_anime)
 
         serializer_data = []
-        serializer_data.extend(API.create_show_objects(self.shows))
+        serializer_data.extend(ShowAPI.create_show_objects(self.shows))
 
         return success_response(serializer_data)