--- a/src/flick/search/tests.py
+++ b/src/flick/search/tests.py
@@ -1,3 +1,0 @@
-from django.test import TestCase
-
-# Create your tests here.