… src/flick/api/urls.py
… src/flick/dump.rdb
… src/flick/rating/migrations/0004_auto_20200802_1803.py
… src/flick/rating/migrations/0005_auto_20200802_1831.py
… src/flick/requirements.txt
… src/flick/search/tests/test_search_shows.py
… src/flick/show/migrations/0003_auto_20200802_0637.py
… src/flick/show/migrations/0004_auto_20200802_0708.py
… src/flick/show/migrations/0005_auto_20200802_0722.py
… src/flick/show/migrations/0006_auto_20200802_0723.py
… src/flick/show/serializers.py
… src/flick/show/show_api_utils.py
… src/flick/show/simple_serializers.py
… src/flick/show/tests/test_show_rating.py
… src/flick/show/tmdb_api_utils.py
… src/flick/show/utils.py
… src/flick/show/views.py
… src/flick/tag/migrations/0003_auto_20200802_1954.py
… src/flick/tag/models.py
… src/flick/tag/serializers.py
… src/flick/tag/simple_serializers.py
… src/flick/tag/views.py
… src/flick/upload/utils.py
… src/flick/upload/views.py