## Overview
@ViviYe @olivialy524 Action item is for you to: `pip install django-slowtests` once this PR is merged -- then you'll get to see something like this:
```
(venv) ~/D/f/s/flick ❯❯❯ python manage.py test
Creating test database for alias 'default'...
System check identified no issues (0 silenced).
........
----------------------------------------------------------------------
Ran 9 tests in 3.866s

OK
Destroying test database for alias 'default'...

9 slowest tests:
1.4452s test_search_show (search.tests.test_search_shows.SearchShowsTests)
0.7901s test_friends_rating (show.tests.test_show_rating.ShowRatingsTests)
0.7317s test_user_can_rate_show (show.tests.test_show_rating.ShowRatingsTests)
0.2638s test_me (flick_auth.tests.test_me.MeTests)
0.2532s test_login (flick_auth.tests.test_login.LoginTests)
0.1134s test_invalid_login (flick_auth.tests.test_login.LoginTests)
0.0908s test_username_is_generated (flick_auth.tests.test_register.RegisterTests)
0.0882s test_missing_social_id_token (flick_auth.tests.test_register.RegisterTests)
0.0881s test_register (flick_auth.tests.test_register.RegisterTests)
```
This way we can start keeping track of what changes are actually optimizing the endpoint and which are slowing it down.



## Changes Made
Add test performance times and add the search test



## Test Coverage
All tests pass (shown above)


## Next Steps
Optimize more lmao